Remove NciIndex trait impl for u32 & add NciIndex trait impl for usize

+++ non_contiguously_indexed_array/tests/constants.rs
@@ -1,11 +1,11 @@
 use non_contiguously_indexed_array::NciArray;
 
-pub const ARRAY_1: NciArray<u32, u32> = NciArray {
+pub const ARRAY_1: NciArray<usize, usize> = NciArray {
     index_range_starting_indices: &[0, 10, 100],
     index_range_skip_amounts: &[0, 7, 95],
     values: &[0, 1, 2, 10, 11, 100],
 };
-pub const ARRAY_2: NciArray<u32, u32> = NciArray {
+pub const ARRAY_2: NciArray<usize, usize> = NciArray {
     index_range_starting_indices: &[100, 200, 500],
     index_range_skip_amounts: &[100, 198, 497],
     values: &[100, 101, 200, 500, 501, 502],
